+++ botan-low/test/Test/Botan/Low/BlockCipher.hs
@@ -1,12 +1,25 @@
-module Main (main) where
+{-# LANGUAGE OverloadedStrings #-}
 
-import           Test.Prelude
+module Test.Botan.Low.BlockCipher (
+    tests
+  ) where
 
 import           Botan.Low.BlockCipher
+import           Botan.Low.Hash
 import           Botan.Low.RNG
+import           Test.Prelude
+import           Test.Tasty
+import           Test.Tasty.Hspec
 
-main :: IO ()
-main = hspec $ testSuite allBlockCiphers chars $ \ bc -> do
+tests :: IO TestTree
+tests = do
+    specs <- testSpec "spec_blockCipher" spec_blockCipher
+    pure $ testGroup "Test.Botan.Low.BlockCipher" [
+        specs
+      ]
+
+spec_blockCipher :: Spec
+spec_blockCipher = testSuite allTestBlockCiphers chars $ \ bc -> do
     it "can initialize a block cipher context" $ do
         _ctx <- blockCipherInit bc
         pass
@@ -61,3 +74,10 @@ main = hspec $ testSuite allBlockCiphers chars $ \ bc -> do
         decmsg <- blockCipherDecryptBlocks ctx encmsg
         decmsg `shouldBe` msg
         pass
+
+allTestBlockCiphers :: [BlockCipherName]
+allTestBlockCiphers = fmap adjust allBlockCiphers
+  where
+    adjust Cascade = cascade Serpent AES256
+    adjust Lion    = lion SHA1 "RC4" (Just 64)
+    adjust name    = name